Cleanup and moving things around
