+++ js/binding.gyp
@@ -2,7 +2,7 @@
   "targets": [
     {
       "target_name": "erlpackjs",
-      "sources": [ "encoder.cc" ],
+      "sources": [ "erlpack.cc" ],
       "include_dirs": [
         "<!(node -e \"require('nan')\")"
       ]
